requirements.txt
flask==3.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-dateutil==2.8.2
email-validator==2.1.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0 